{"schema":"decodex/commit/1","summary":"Split SQLite state store modules","authority":"XY-1115"}
